requirements.txt
anthropic>=0.18.0
python-dotenv>=1.0.0
flask>=2.3.0
werkzeug>=2.3.0
python-docx>=0.8.11
gunicorn>=20.1.0
